Move config package into the pkg directory
